refactor: 컴포넌트, style 분리
